reverse di una lista con int

+++ liste/liste_collegate/reverse/lista.h
@@ -1,4 +1,4 @@
-typedef float Dato;
+typedef int Dato;
 
 typedef struct nodo
 {
